requirements.txt
apscheduler
git+https://github.com/plamere/spotipy.git
Pillow
packaging
paho-mqtt
click